[DEBUG] partion file instead of dir
